# contracts/openapi/control-plane.v1.yaml
openapi: 3.1.0
...
info:
  title: Talosmine Control Plane API
  version: 1.0.0
  description: |
    Control Plane của Talosmine — nguồn sự thật về identity, entitlement và quota.

    **Trạng thái: P1 (bootstrap).** Chỉ có health endpoint. Không có capability nghiệp vụ nào.

    Quy ước xuyên phase:
    - Số lượng (quantity/limit/counter/amount) biểu diễn bằng **decimal string**, không phải
      JSON number — PostgreSQL dùng `bigint` và JSON number mất chính xác ngoài 2^53.
    - Thời gian dùng timestamp có timezone; business time dựa DB clock.
    - Mọi lỗi trả về `ErrorEnvelope` với `code` máy đọc được. Client bắt theo `code`,
      không theo `message`.
  license:
    name: UNLICENSED
servers:
  - url: https://talosmine.localhost
    description: Local dev qua Caddy (DEC-T11)
...
paths:
...
  /v1/admin/survey/questions:
    get:
      tags: [admin]
      operationId: adminListSurveyQuestions
      summary: Câu hỏi khảo sát (quản trị)
      description: |
        Yêu cầu `content:read`. Trả MỌI trạng thái và MỌI ngôn ngữ — màn hình quản trị cần
        thấy cả những gì người dùng chưa thấy, và sửa được cả hai bản dịch cạnh nhau.

        KHÔNG có endpoint tạo/xoá câu hỏi: cấu trúc ba câu là cố định, seed bằng migration
        vì code phải có chỗ render tương ứng. Chỉ nội dung và danh sách lựa chọn là mở.
      security:
        - webSession: []
      responses:
        '200':
          description: Danh sách câu hỏi, sắp theo `sortOrder`.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminSurveyQuestion'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /v1/admin/survey/questions/{questionId}:
    patch:
      tags: [admin]
      operationId: adminUpdateSurveyQuestion
      summary: Sửa nội dung một câu hỏi
      description: |
        Yêu cầu `content:manage`. Sửa được `minSelect` và bản dịch.

        KHÔNG sửa được `key` (code tham chiếu nó để render đúng layout) và `kind` (đổi
        `multi` thành `single` sẽ làm dữ liệu đã thu thập mâu thuẫn với ràng buộc mới).
      security:
        - webSession: []
      parameters:
        - $ref: '#/components/parameters/SurveyQuestionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSurveyQuestionRequest'
      responses:
        '204':
          description: Đã lưu.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/admin/survey/options:
    post:
      tags: [admin]
      operationId: adminCreateSurveyOption
      summary: Thêm một lựa chọn
      description: |
        Yêu cầu `content:manage`. Lựa chọn mới LUÔN ở `draft` — đưa ra trước người dùng là
        hành động riêng cần `content:publish`.

        `sortOrder` bỏ trống thì xếp cuối. `icon` phải thuộc danh mục đóng.
      security:
        - webSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSurveyOptionRequest'
      responses:
        '201':
          description: Đã tạo.
          content:
            application/json:
              schema:
                type: object
                required: [id]
                additionalProperties: false
                properties:
                  id: { type: string, format: uuid }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: '`key` đã tồn tại trong câu hỏi này.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /v1/admin/survey/options/reorder:
    post:
      tags: [admin]
      operationId: adminReorderSurveyOptions
      summary: Sắp xếp lại lựa chọn trong một câu hỏi
      description: |
        Yêu cầu `content:manage`. Nhận TOÀN BỘ id của câu hỏi theo thứ tự mới — cùng lý do
        với `/v1/admin/site/nav/reorder`: kết quả không phụ thuộc thứ tự request tới.
      security:
        - webSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReorderSurveyOptionsRequest'
      responses:
        '204':
          description: Đã sắp xếp lại.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /v1/admin/survey/options/{optionId}:
    patch:
      tags: [admin]
      operationId: adminUpdateSurveyOption
      summary: Sửa icon và nhãn của một lựa chọn
      description: |
        Yêu cầu `content:manage`. KHÔNG sửa được `key` — câu trả lời đã thu thập tham chiếu
        lựa chọn này, và đổi khoá sẽ làm mọi số liệu lịch sử trỏ sai đối tượng.
      security:
        - webSession: []
      parameters:
        - $ref: '#/components/parameters/SurveyOptionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSurveyOptionRequest'
      responses:
        '204':
          description: Đã lưu.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [admin]
      operationId: adminDeleteSurveyOption
      summary: Xoá một lựa chọn
      description: |
        Yêu cầu `content:manage`.

        Lựa chọn ĐÃ CÓ NGƯỜI TRẢ LỜI thì KHÔNG xoá được — khoá ngoại `ON DELETE RESTRICT` ở
        `survey_answers` chặn, và API trả 409. Đó là hành vi đúng: xoá nó sẽ làm mọi số liệu
        lịch sử trỏ vào hư không. Muốn gỡ khỏi giao diện thì đổi trạng thái sang `inactive`.
      security:
        - webSession: []
      parameters:
        - $ref: '#/components/parameters/SurveyOptionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminReason'
      responses:
        '204':
          description: Đã xoá.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Lựa chọn đã có người trả lời — dùng `inactive` thay vì xoá.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /v1/admin/survey/options/{optionId}/status:
    post:
      tags: [admin]
      operationId: adminChangeSurveyOptionStatus
      summary: Đưa lựa chọn ra hoặc gỡ khỏi khảo sát
      description: |
        Yêu cầu `content:publish` — permission RIÊNG, không phải `content:manage`. Cùng lập
        luận với điều hướng site: đưa một lựa chọn sang `active` là đặt nó trước mặt mọi
        người dùng mới đăng ký.

        Chỉ nhận `active` và `inactive`; `draft` không phải đích hợp lệ của chuyển tiếp nào.
      security:
        - webSession: []
      parameters:
        - $ref: '#/components/parameters/SurveyOptionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeContentStatusRequest'
      responses:
        '204':
          description: Đã đổi trạng thái.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  responses:
    BadRequest:
      description: Dữ liệu vào không hợp lệ (ví dụ thiếu `reason`).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unauthorized:
      description: Chưa xác thực (thiếu/sai/hết hạn/đã thu hồi phiên).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Forbidden:
      description: |
        Đã xác thực nhưng KHÔNG đủ quyền quản trị. Cố ý không nêu thiếu permission nào —
        tránh vẽ bản đồ phân quyền cho kẻ dò.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFound:
      description: Không tìm thấy tài nguyên trong phạm vi được phép.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  securitySchemes:
    webSession:
      type: apiKey
      in: header
      name: X-Session-Token
      description: |
        Web session token do BFF forward (từ cookie HttpOnly). Control Plane tra hash
        trong `web_sessions`, kiểm chưa hết hạn/chưa thu hồi, rồi resolve accountId.
        Token thô KHÔNG được log. Đây là kênh của NGƯỜI DÙNG — khác `serviceIdentity` (M2M).
    userSession:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token user do Auth0 phát, được BFF chuyển tiếp.
        **CHƯA dùng ở P1** — wiring Auth0 thật là P2 và phụ thuộc DEC-B03 (tenant thật).
    serviceIdentity:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        M2M token của backend ứng dụng (Data Plane). Mỗi backend có identity RIÊNG,
        audience cụ thể và exact resource scope. Không dùng chung credential.
        **CHƯA dùng ở P1.**
  schemas:
    AdminReason:
      type: object
      description: |
        Mọi mutation quản trị BẮT BUỘC có lý do (modular.md mục 11.4). Lý do đi vào audit
        và là thứ khiến bản ghi audit có giá trị điều tra sau này.
      required: [reason]
      additionalProperties: false
      properties:
        reason:
          type: string
          minLength: 1
    # ── Survey (onboarding) ────────────────────────────────────────────────────
    SurveyQuestionKey:
      type: string
      enum: [categories, primary_use, discover_first]
      description: |
        Khoá máy ỔN ĐỊNH. Danh mục ĐÓNG — cấu trúc khảo sát cố định vì code phải có chỗ
        render tương ứng; chỉ nội dung và danh sách lựa chọn sửa được trong `/admin`.
    SurveyIconKey:
      type: string
      description: |
        Khoá icon thuộc danh mục ĐÓNG; code render SVG tương ứng. Không phải URL và không
        phải SVG — CSP theo nonce (DEC-T20) sẽ chặn markup tự nhập, và nới CSP để chiều một
        ô nhập icon 20px là đánh đổi tệ nhất trong kiến trúc này.

        Danh sách này phải khớp CHECK ở migration `0012_survey` VÀ `survey-icons.tsx`.
      enum:
        [
          writing,
          design,
          code,
          video,
          image,
          automation,
          research,
          business,
          chart,
          chat,
          cloud,
          shield,
          sparkle,
          rocket,
          book,
        ]
...
    AdminSurveyQuestion:
      type: object
      required:
        [id, key, kind, minSelect, sortOrder, status, titles, descriptions, options, updatedAt]
      additionalProperties: false
      description: |
        Khác `SurveyQuestion` ở hai chỗ: trả CẢ HAI ngôn ngữ cùng lúc (màn hình quản trị sửa
        song song), và trả cả lựa chọn `draft`/`inactive` mà người dùng không thấy.
      properties:
        id: { type: string, format: uuid }
        key:
          $ref: '#/components/schemas/SurveyQuestionKey'
        kind: { type: string, enum: [single, multi] }
        minSelect: { type: integer, minimum: 1 }
        sortOrder: { type: integer, minimum: 0 }
        status: { type: string, enum: [active, inactive] }
        titles:
          $ref: '#/components/schemas/LocalizedText'
        descriptions:
          $ref: '#/components/schemas/LocalizedText'
        options:
          type: array
          description: Đã sắp theo `sortOrder`, gồm MỌI trạng thái.
          items:
            $ref: '#/components/schemas/AdminSurveyOption'
        updatedAt: { type: string, format: date-time }
    AdminSurveyOption:
      type: object
      required: [id, key, icon, sortOrder, status, labels, descriptions, updatedAt]
      additionalProperties: false
      properties:
        id: { type: string, format: uuid }
        key:
          type: string
          description: Ổn định trong phạm vi câu hỏi. Không sửa được sau khi tạo.
        icon:
          oneOf:
            - $ref: '#/components/schemas/SurveyIconKey'
            - type: 'null'
        sortOrder: { type: integer, minimum: 0 }
        status: { type: string, enum: [draft, active, inactive] }
        labels:
          $ref: '#/components/schemas/LocalizedText'
        descriptions:
          $ref: '#/components/schemas/LocalizedText'
        updatedAt: { type: string, format: date-time }
    UpdateSurveyQuestionRequest:
      type: object
      required: [reason]
      additionalProperties: false
      description: |
        Trường vắng mặt = không đổi. `titles.vi = null` là XOÁ bản dịch tiếng Việt, khác với
        việc không gửi `titles` (giữ nguyên).
      properties:
        minSelect:
          type: integer
          minimum: 1
          description: |
            Chỉ có nghĩa với câu `multi`. Đặt lớn hơn số lựa chọn `active` hiện có sẽ bị từ
            chối — nếu không, không ai nộp nổi khảo sát.
        titles:
          $ref: '#/components/schemas/LocalizedText'
        descriptions:
          $ref: '#/components/schemas/LocalizedText'
        reason: { type: string, minLength: 1 }
    CreateSurveyOptionRequest:
      type: object
      required: [questionKey, key, labels, reason]
      additionalProperties: false
      properties:
        questionKey:
          $ref: '#/components/schemas/SurveyQuestionKey'
        key:
          type: string
          minLength: 1
          maxLength: 64
          pattern: '^[a-z0-9_]+$'
          description: |
            Chữ thường, số và gạch dưới. Ràng buộc hình dạng vì đây là khoá máy đi vào dữ
            liệu lịch sử và báo cáo, không phải chữ hiển thị.
        icon:
          oneOf:
            - $ref: '#/components/schemas/SurveyIconKey'
            - type: 'null'
        labels:
          $ref: '#/components/schemas/LocalizedText'
        descriptions:
          $ref: '#/components/schemas/LocalizedText'
        sortOrder:
          type: integer
          minimum: 0
          description: Bỏ trống thì xếp cuối câu hỏi.
        reason: { type: string, minLength: 1 }
    UpdateSurveyOptionRequest:
      type: object
      required: [reason]
      additionalProperties: false
      description: 'Trường vắng mặt = không đổi. `key` KHÔNG sửa được — xem mô tả endpoint.'
      properties:
        icon:
          oneOf:
            - $ref: '#/components/schemas/SurveyIconKey'
            - type: 'null'
        labels:
          $ref: '#/components/schemas/LocalizedText'
        descriptions:
          $ref: '#/components/schemas/LocalizedText'
        reason: { type: string, minLength: 1 }
    ReorderSurveyOptionsRequest:
      type: object
      required: [questionKey, optionIds, reason]
      additionalProperties: false
      properties:
        questionKey:
          $ref: '#/components/schemas/SurveyQuestionKey'
        optionIds:
          type: array
          minItems: 1
          description: TOÀN BỘ id của câu hỏi, theo thứ tự mới. Thiếu hoặc thừa đều bị từ chối.
          items: { type: string, format: uuid }
        reason: { type: string, minLength: 1 }
...
    LocalizedText:
      type: object
      description: |
        Một đoạn văn bản theo từng ngôn ngữ. Dùng chung cho điều hướng site và khảo sát —
        cùng một hình dạng, cùng một quy tắc, nên không tách làm hai.

        Thiếu một ngôn ngữ là hợp lệ ở trạng thái `draft` — nhưng mục đó sẽ bị BỎ QUA ở
        đường công khai của ngôn ngữ còn thiếu (join INNER, không fallback).
      additionalProperties: false
      properties:
        vi: { type: [string, 'null'] }
        en: { type: [string, 'null'] }
...
    ChangeContentStatusRequest:
      type: object
      required: [status, reason]
      additionalProperties: false
      properties:
        status:
          type: string
          enum: [active, inactive]
          description: '`draft` không phải đích hợp lệ — xem mô tả endpoint.'
        reason: { type: string, minLength: 1 }
...
    ErrorEnvelope:
      type: object
      description: |
        Hình dạng lỗi ổn định cho MỌI lỗi rời khỏi API, kể cả lỗi không lường trước.

        Ràng buộc bảo mật: không bao giờ chứa stack trace, secret, connection string
        hay chi tiết nội bộ. Với lỗi 5xx, `message` luôn là thông điệp chung — chi tiết
        thật chỉ đi vào log phía server, truy được bằng `correlationId`.
      required: [code, message, correlationId]
      additionalProperties: false
      properties:
        code:
          type: string
          description: |
            Mã máy đọc được, ổn định, không phụ thuộc câu chữ giao diện.
            Client bắt lỗi theo trường này.
          pattern: '^[A-Z][A-Z0-9_]*$'
          examples:
            - INTERNAL_ERROR
            - VALIDATION_FAILED
            - SERVICE_UNAVAILABLE
        message:
          type: string
          description: Thông điệp an toàn để hiển thị.
          maxLength: 500
        correlationId:
          type: string
          format: uuid
          description: |
            Truy vết xuyên BFF -> API -> worker -> log.
            Nhận từ header `X-Correlation-Id` nếu hợp lệ, ngược lại server sinh mới (UUIDv7).
            **Không phải** idempotency key — đừng dùng lẫn.
        details:
          type: object
          description: Chi tiết bổ sung theo allowlist. Không bao giờ chứa dữ liệu nhạy cảm.
          additionalProperties: true
  parameters:
...
    SurveyQuestionId:
      name: questionId
      in: path
      required: true
      schema: { type: string, format: uuid }
    SurveyOptionId:
      name: optionId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
security: []